pyproject.toml
[tool.poetry]
name = "blog"
version = "0.1.0"
description = ""
authors = ["macadden <mati.macadden@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.6.0"
psycopg2-binary = "^2.9.10"
djangorestframework-simplejwt = "^5.3.1"
django-simple-history = "^3.7.0"
django-filter = "^24.3"
django-environ = "^0.11.2"
pyjwt = "^2.9.0"
social-auth-app-django = "^5.4.2"
social-auth-core = "^4.5.4"
django-storages = "^1.14.4"
djoser = "^2.3.1"
argon2-cffi = "^23.1.0"
pillow = "^11.0.0"
python-slugify = "^8.0.4"
django-ckeditor = "^6.7.1"
python-dotenv = "^1.0.1"
celery = "^5.4.0"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
mixer = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
